feat(template-basic): update plugin structure and dependencies
- Refactored the template-basic project to use Preact instead of React, updating the Vite configuration and dependencies accordingly.
- Added new dependencies for Preact and updated the TypeScript configuration to support the new JSX settings.
- Replaced the old index.ts file with a new index.tsx file that implements the ExamplePlugin with a toolbar action.
- Cleaned up the package.json files to reflect the changes in dependencies and structure.
