+++ packages/create-stagewise-plugin/package.json
@@ -8,7 +8,11 @@
     "create-stagewise-plugin": "index.js",
     "cva": "index.js"
   },
-  "files": ["index.js", "template-*", "dist"],
+  "files": [
+    "index.js",
+    "template-*",
+    "dist"
+  ],
   "scripts": {
     "dev": "unbuild --stub",
     "build": "unbuild",
@@ -35,5 +39,9 @@
     "mri": "^1.2.0",
     "picocolors": "^1.1.1",
     "unbuild": "^3.5.0"
+  },
+  "dependencies": {
+    "@types/react": "^19.1.3",
+    "react": "^19.1.0"
   }
 }
