+++ packages/create-stagewise-plugin/template-basic/package.json
@@ -27,15 +27,17 @@
     "build": "tsc -b && vite build --mode production"
   },
   "dependencies": {
-    "axe-core": "^4.10.3"
+    "@stagewise/plugin-a11y": "link:",
+    "axe-core": "^4.10.3",
+    "preact": "^10.26.6"
   },
   "devDependencies": {
+    "@preact/preset-vite": "^2.10.1",
     "@stagewise/toolbar": "workspace:*",
+    "rollup-preserve-directives": "^1.0.0",
     "typescript": "~5.8.3",
     "vite": "^6.3.5",
-    "vite-plugin-dts": "^4.5.3",
-    "@vitejs/plugin-react-swc": "^3.7.0",
-    "rollup-preserve-directives": "^1.0.0"
+    "vite-plugin-dts": "^4.5.3"
   },
   "packageManager": "pnpm@10.10.0"
 }
